+++ internal/handlers/metric_test.go
@@ -1,18 +1,16 @@
 package handlers
 
 import (
-	"fmt"
-	"github.com/mvigor/metricsd/cmd/server/storage"
+	"github.com/mvigor/metricsd/internal/storage"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 	"io"
 	"net/http"
 	"net/http/httptest"
-	"sort"
 	"testing"
 )
 
-func TestIndexHandler(t *testing.T) {
+func TestShowHandler(t *testing.T) {
 	storage := storage.MemoryStorage{
 		Metrics: map[string]string{
 			"metric1": "1",
@@ -22,21 +20,10 @@ func TestIndexHandler(t *testing.T) {
 		},
 	}
 
-	successBody := ""
-	keys := make([]string, 0, len(storage.Metrics))
-	for k := range storage.Metrics {
-		keys = append(keys, k)
-	}
-	sort.Strings(keys)
-
-	for _, k := range keys {
-		val := storage.Metrics[k]
-		successBody += fmt.Sprintf("%s = %s<br>\n", k, val)
-	}
-
 	testCases := []struct {
 		name                string
 		method              string
+		request             map[string]string
 		expectedCode        int
 		expectedBody        string
 		expectedContentType string
@@ -44,25 +31,45 @@ func TestIndexHandler(t *testing.T) {
 		{
 			name:                "test case #1",
 			method:              http.MethodGet,
+			request:             map[string]string{"metric_name": "metric1"},
 			expectedCode:        http.StatusOK,
-			expectedBody:        successBody,
+			expectedBody:        "metric name = metric1, value = 1",
 			expectedContentType: "text/html",
 		},
 		{
-			name:         "test case #2",
-			method:       http.MethodPut,
-			expectedCode: http.StatusMethodNotAllowed,
-			expectedBody: "",
+			name:                "test case #2",
+			method:              http.MethodGet,
+			request:             map[string]string{"metric_name": "metric3"},
+			expectedCode:        http.StatusOK,
+			expectedBody:        "metric name = metric3, value = value2",
+			expectedContentType: "text/html",
 		},
 		{
-			name:         "test case #3",
-			method:       http.MethodDelete,
-			expectedCode: http.StatusMethodNotAllowed,
-			expectedBody: "",
+			name:                "test case #3",
+			method:              http.MethodGet,
+			request:             map[string]string{"metric_name": "metri"},
+			expectedCode:        http.StatusNotFound,
+			expectedBody:        "",
+			expectedContentType: "text/html",
 		},
 		{
 			name:         "test case #4",
+			method:       http.MethodPut,
+			request:      map[string]string{"metric_name": "metric12"},
+			expectedCode: http.StatusMethodNotAllowed,
+			expectedBody: "",
+		},
+		{
+			name:         "test case #5",
+			method:       http.MethodDelete,
+			request:      map[string]string{"metric_name": "metric12"},
+			expectedCode: http.StatusMethodNotAllowed,
+			expectedBody: "",
+		},
+		{
+			name:         "test case #6",
 			method:       http.MethodPost,
+			request:      map[string]string{"metric_name": "metric12"},
 			expectedCode: http.StatusMethodNotAllowed,
 			expectedBody: "",
 		},
@@ -71,7 +78,7 @@ func TestIndexHandler(t *testing.T) {
 		t.Run(tt.name, func(t *testing.T) {
 			request := httptest.NewRequest(tt.method, "/", nil)
 			w := httptest.NewRecorder()
-			handlerFunc := IndexHandler(nil, &storage)
+			handlerFunc := ShowHandler(tt.request, &storage)
 			handlerFunc(w, request)
 			res := w.Result()
 			assert.Equal(t, res.StatusCode, tt.expectedCode)
