+++ internal/storage/memory.go
